Refactor import paths in Login and Register components to simplify structure
