+++ client/src/pages/auth/register/RegisterForm.tsx
